SQL columns are now float

+++ auxiliary_models.py
@@ -10,8 +10,8 @@ class Location(db.Model):
     __tablename__ = 'location'
 
     id = db.Column(db.Integer, primary_key=True)
-    latitude = db.Column(db.Double())
-    longitude = db.Column(db.Double())
+    latitude = db.Column(db.Float())
+    longitude = db.Column(db.Float())
     existent = db.Column(db.Boolean())
 
     person_id = db.Column(db.ForeignKey('person.id'))
